requirements.txt
discord.py==2.6.4
PyNaCl>=1.5,<1.6